Refactoring: moved search condition state into context provider
