+++ src/pages/knowledgegraph/GraphSearch/useSubConditions.ts
@@ -2,7 +2,7 @@ import { useMemo } from 'react';
 import { useDataModel } from '@/store';
 import { W3CAnnotation, W3CAnnotationBody } from '@annotorious/react';
 import { serializePropertyValue } from '@/utils/serialize';
-import { DropdownOption } from './Types';
+import { DropdownOption } from '../Types';
 
 export const useSubConditions = (
   annotations: W3CAnnotation[],
